fix(availability): remove unused example property in availability.yaml feat(availability): add lookup property to availability.yaml feat(availability): add lookup.yaml file fix(availability): add missing required properties in get/body.yaml feat(availability): add availability.types.ts file feat(availability): add index.ts file fix(availability): update mock data in availability.spec.ts fix(lookup): update mock data in lookup.service.spec.ts fix(schedule): remove unused Availability type fix(availability): update import path in find-start-end-date-in-availability.ts fix(availability): update import path in generate-availability.spec.ts fix(availability): update import path in generate-availability.ts fix(availability): update import path in remove-booked-slots.spec.ts fix(availability): update import path in remove-booked-slots.ts
