+++ src/functions/availability/index.ts
@@ -1,0 +1,1 @@
+export * from "./availability.types";
