+++ src/functions/schedule/schedule.types.ts
@@ -5,7 +5,6 @@ import {
   NumberOrStringType,
   StringOrObjectIdType,
 } from "~/library/zod";
-import { User } from "../user";
 
 export enum TimeUnit {
   HOURS = "hours",
@@ -174,21 +173,3 @@ export const ScheduleZodSchema = z.object({
 });
 
 export type Schedule = z.infer<typeof ScheduleZodSchema>;
-
-export type Availability = {
-  date: Date;
-  customer: Pick<User, "fullname" | "customerId">;
-  slots: {
-    from: Date;
-    to: Date;
-    products: {
-      productId: number;
-      variantId: number;
-      from: Date;
-      to: Date;
-      breakTime: number;
-      duration: number;
-      travelTime: number;
-    }[];
-  }[];
-};
